pyproject.toml
[tool.poetry]
name = "tg-delete-system-messages"
version = "0.1.0"
description = ""
authors = ["Ilia Kaziamov <kaziamov@outlook.com>"]
readme = "README.md"
packages = [{include = "tg-cleaner"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.2"
aiogram = "^3.1.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
